classification_engine/requirements.txt
filelock==3.20.3
fsspec==2026.1.0
grpcio==1.76.0
PyJWT==2.10.1
grpcio-tools==1.76.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.2
packaging==26.0
pillow==12.1.0
protobuf==6.33.5
python-dotenv==1.0.0
requests==2.31.0
setuptools==80.10.2
sympy==1.14.0
torch==2.10.0      #+cu121          # Tommy - Torch version that support GPU
torchvision==0.25.0      #+cu121   # Tommy - Torchvision version that support GPU
typing_extensions==4.15.0
wheel==0.46.3
#nvidia-ml-py3==11.495.46    # Tommy - Add NVIDIA-ml-py3 for GPU  --Removed, Needed highly specific hardware I don't have despite having CUDA
psutil==5.9.5               # Tommy - Add PSUtil for CPU power calculation
timm==1.0.26                # Tommy - Add Timm so we can use ImageNet 20k (20k images) instead of 1k
